module_27_postgres_migrations/hw/requirements.txt
Flask==2.0.3
gunicorn==20.1.0
psycopg2-binary==2.9.1
werkzeug>=2,<3
flask-sqlalchemy
requests